Move the interfaces to the relevant files
