+++ lib/device.ts
@@ -3,10 +3,7 @@
  *
  * SPDX-License-Identifier: MIT
  */
-import { AccelerometerData } from "./accelerometer.js";
 import { TypedEventTarget } from "./events.js";
-import { LedMatrix } from "./led.js";
-import { MagnetometerData } from "./magnetometer.js";
 import { UARTDataEvent } from "./uart.js";
 
 /**
@@ -238,152 +235,4 @@ export interface DeviceConnection
    * Clear device to enable chooseDevice.
    */
   clearDevice(): Promise<void> | void;
-}
-
-export interface MicrobitWebUSBConnection extends DeviceConnection {
-  /**
-   * Gets micro:bit deviceId.
-   *
-   * @returns the device id or undefined if there is no connection.
-   */
-  getDeviceId(): number | undefined;
-
-  /**
-   * Sets device request exclusion filters.
-   */
-  setRequestDeviceExclusionFilters(exclusionFilters: USBDeviceFilter[]): void;
-
-  /**
-   * Flash the micro:bit.
-   *
-   * @param dataSource The data to use.
-   * @param options Flash options and progress callback.
-   */
-  flash(dataSource: FlashDataSource, options: {}): Promise<void>;
-
-  /**
-   * Gets micro:bit device.
-   *
-   * @returns the USB device or undefined if there is no connection.
-   */
-  getDevice(): USBDevice | undefined;
-
-  /**
-   * Resets the micro:bit in software.
-   */
-  softwareReset(): Promise<void>;
-}
-
-export interface MicrobitWebBluetoothConnection extends DeviceConnection {
-  /**
-   * Sets micro:bit name filter for device requesting.
-   *
-   * @param name The name of the micro:bit.
-   */
-  setNameFilter(name: string): void;
-
-  /**
-   * Gets micro:bit accelerometer data.
-   *
-   * @returns accelerometer data or undefined if there is no connection.
-   */
-  getAccelerometerData(): Promise<AccelerometerData | undefined>;
-
-  /**
-   * Gets micro:bit accelerometer period.
-   *
-   * @returns accelerometer period or undefined if there is no connection.
-   */
-  getAccelerometerPeriod(): Promise<number | undefined>;
-
-  /**
-   * Sets micro:bit accelerometer period.
-   *
-   * @param value The accelerometer period.
-   */
-  setAccelerometerPeriod(value: number): Promise<void>;
-
-  /**
-   * Sets micro:bit LED text.
-   *
-   * @param text The text displayed on micro:bit LED.
-   */
-  setLedText(text: string): Promise<void>;
-
-  /**
-   * Gets micro:bit LED scrolling delay.
-   *
-   * @returns LED scrolling delay in milliseconds.
-   */
-  getLedScrollingDelay(): Promise<number | undefined>;
-
-  /**
-   * Sets micro:bit LED scrolling delay.
-   *
-   * @param delayInMillis LED scrolling delay in milliseconds.
-   */
-  setLedScrollingDelay(delayInMillis: number): Promise<void>;
-
-  /**
-   * Gets micro:bit LED matrix.
-   *
-   * @returns a boolean matrix representing the micro:bit LED display.
-   */
-  getLedMatrix(): Promise<LedMatrix | undefined>;
-
-  /**
-   * Sets micro:bit LED matrix.
-   *
-   * @param matrix an boolean matrix representing the micro:bit LED display.
-   */
-  setLedMatrix(matrix: LedMatrix): Promise<void>;
-
-  /**
-   * Gets micro:bit magnetometer data.
-   *
-   * @returns magnetometer data.
-   */
-  getMagnetometerData(): Promise<MagnetometerData | undefined>;
-
-  /**
-   * Gets micro:bit magnetometer bearing.
-   *
-   * @returns magnetometer bearing.
-   */
-  getMagnetometerBearing(): Promise<number | undefined>;
-
-  /**
-   * Gets micro:bit magnetometer period.
-   *
-   * @returns magnetometer period.
-   */
-  getMagnetometerPeriod(): Promise<number | undefined>;
-
-  /**
-   * Sets micro:bit magnetometer period.
-   *
-   * @param value magnetometer period.
-   */
-  setMagnetometerPeriod(value: number): Promise<void>;
-
-  /**
-   * Triggers micro:bit magnetometer calibration.
-   */
-  triggerMagnetometerCalibration(): Promise<void>;
-
-  /**
-   * Write UART messages.
-   *
-   * @param data UART message.
-   */
-  writeUART(data: Uint8Array): Promise<void>;
-}
-
-export interface MicrobitRadioBridgeConnection extends DeviceConnection {
-  /**
-   * Sets remote device.
-   *
-   * @param deviceId The device id of remote micro:bit.
-   */
-  setRemoteDeviceId(deviceId: number): void;
 }
